pyproject.toml
[project]
name = "weave-evals-traces"
version = "0.1.0"
description = "Pharmaceutical QA evaluation examples using Weave"
dependencies = [
    "weave>=0.51.44",
    "openai>=1.0.0",
    "anthropic>=0.34.0",
    "jinja2>=3.1.0",
    "python-dotenv>=1.0.0",
    "pillow>=10.0.0",
]
requires-python = ">= 3.9"

[tool.setuptools]
py-modules = []

[tool.ruff]
line-length = 88
target-version = "py39"